Refactor: Move external function definitions into a separate file
